requirements.txt
uvicorn[standard]
fastapi==0.110.0
passlib==1.7.4
types-passlib==1.7.7.20240106
bcrypt==4.0.1
pytest==8.0.2
# pytest-asyncio==0.23.5
# pytest-postgresql==5.1.0
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
alembic==1.13.1
asyncpg==0.29.0
mypy==1.8.0
sqlalchemy[mypy]==2.0.27
pytz==2024.1
types-pytz==2024.1.0.20240203